confirm popup v1.0.0

+++ src/app/quizzes/quizz-card/quizz-card.component.ts
@@ -6,6 +6,7 @@ import { QuizzesService } from '../services/quizzes.service';
 import { Store } from '@ngrx/store';
 import { AppState } from 'src/app/reducers';
 import { DeleteQuiz } from '../store/quizzes.actions';
+import { ConfirmPopupService } from 'src/app/core/services/confirm-popup.service';
 
 @Component({
   selector: 'app-quiz-card',
@@ -13,11 +14,12 @@ import { DeleteQuiz } from '../store/quizzes.actions';
   styleUrls: ['./quiz-card.component.scss']
 })
 export class QuizCardComponent implements OnInit {
-  
+
   @Input() quiz;
   author$: Observable<any>;
 
-  constructor(private authService: AuthService,private store: Store<AppState>, private quizService: QuizzesService) { }
+  constructor(private authService: AuthService, private confirmPopupService: ConfirmPopupService,
+              private store: Store<AppState>, private quizService: QuizzesService) { }
 
   ngOnInit() {
     this.author$ = this.authService.getUserById(this.quiz.author).pipe(
@@ -26,7 +28,12 @@ export class QuizCardComponent implements OnInit {
   }
 
   deleteQuiz(e: Event) {
-    this.store.dispatch(new DeleteQuiz({quizId: this.quiz.id}));
+    this.confirmPopupService.confirm({
+      title: 'Delete quiz',
+      message: 'Do you want to delete this quiz?'
+    }).subscribe((confirmed: boolean) => {
+      this.store.dispatch(new DeleteQuiz({quizId: this.quiz.id}));
+    });
   }
 
 }
